+++ infa/dependencies.tf
@@ -1,4 +1,4 @@
-# Fresh VPS for App
+# Fresh VPC for App
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
   version = "2.77.0"
@@ -12,12 +12,12 @@ module "vpc" {
   # Required for private EKS API access
   enable_dns_hostnames = true
 
-  enable_nat_gateway = true
+  enable_nat_gateway = false
 
 }
 
-# Endpoints to allow communicating with the AWS services from ECS soley through private traffic
-# Note: This is currently not used
+# Endpoints to allow communicating with the AWS services from ECS solely through private traffic
+# Note: These endpoints result in not needing the costly NAT Gateway
 module "endpoints" {
   source             = "terraform-aws-modules/vpc/aws//modules/vpc-endpoints"
   vpc_id             = module.vpc.vpc_id
@@ -30,7 +30,7 @@ module "endpoints" {
       service_type    = "Gateway"
       route_table_ids = module.vpc.private_route_table_ids
       tags = {
-        Name = "${var.resource_prefix}-s3-vpc-endpoint"
+        Name = "${var.resource_prefix}s3-vpc-endpoint"
       }
     },
     ecs = {
@@ -47,6 +47,22 @@ module "endpoints" {
       subnet_ids          = module.vpc.private_subnets
       tags = {
         Name = "${var.resource_prefix}logs-vpc-endpoint"
+      }
+    },
+    ecr_api = {
+      service             = "ecr.api"
+      private_dns_enabled = true
+      subnet_ids          = module.vpc.private_subnets
+      tags = {
+        Name = "${var.resource_prefix}ecr-api-vpc-endpoint"
+      }
+    },
+    ecr_dkr = {
+      service             = "ecr.dkr"
+      private_dns_enabled = true
+      subnet_ids          = module.vpc.private_subnets
+      tags = {
+        Name = "${var.resource_prefix}ecr-dkr-vpc-endpoint"
       }
     }
   }