+++ infa/main.tf
@@ -18,14 +18,17 @@ terraform {
 }
 
 provider "aws" {
-  region     = local.region
-  profile    = "personal"
+  region  = local.region
+  profile = "personal"
 }
 
 # General Locals
 locals {
-  region          = "us-east-1"
-  container_port  = 5050
+  region            = "us-east-1"
+  container_port    = 5050
+  enable_scheduling = true
+  start_time        = "cron(30 11 ? * Mon-Fri *)"
+  stop_time         = "cron(0 19 ? * Mon-Fri *)"
 }
 
 # Definition for each task to be ran by ECS Fargate